add new update Blue-Line-OE-4ATV_Skins-for-openATV-python3

+++ Blue-Line-OE-4ATV/usr/lib/enigma2/python/Components/Renderer/BLWatches.py
@@ -1,4 +1,3 @@
-# Embedded file name: /usr/lib/enigma2/python/Components/Renderer/AXBlueWatches.py
 import math
 from Components.Renderer.Renderer import Renderer
 from skin import parseColor
@@ -39,7 +38,7 @@ class BLWatches(Renderer):
     def hand(self):
         width = self.instance.size().width()
         height = self.instance.size().height()
-        r = min(width, height) / 2
+        r = int(min(width, height) / 2)
         endX, endY = self.calculate(self.numval, r, r)
         self.draw_line(r, r, endX, endY)
